pyproject.toml
[tool.poetry]
name = "aitk-559-tagging-of-legal-acts-prototype"
version = "0.1.0"
description = ""
authors = ["nichnikov <a.nichnikov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pypdf2 = "^3.0.1"
pdfminer = "^20191125"
pdfplumber = "^0.11.2"
pillow = "^10.4.0"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
requests = "^2.32.3"
pyarrow = "^16.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
